% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{Combine A biomarkers}
\usage{
A(
  ab40 = bmk(),
  ab42 = bmk(),
  ab42_ab40_ratio = bmk(),
  amyloid_pet = bmk(),
  .classification = c("2018", "2024")
)
}
\arguments{
\item{ab40}{AB40 as a \code{bmk} or character vector (no default; can omit)}

\item{ab42}{AB42 as a \code{bmk} or character vector (no default; can omit)}

\item{ab42_ab40_ratio}{AB42/AB40 ratio as a \code{bmk} or character vector (no default; can omit)}

\item{amyloid_pet}{Amyloid PET as a \code{bmk} or character vector (no default; can omit)}

\item{.classification}{ATN classification scheme.}
}
\value{
A \code{bmk} vector representing the combined A-status --- one biomarker
positive is sufficient to render A-status positive.
}
\description{
Computes the overall A-status by OR-combining any subset of these biomarkers:
\itemize{
\item \code{ab40}
\item \code{ab42}
\item \code{ab42_ab40_ratio}
\item \code{amyloid_pet}
}

At least one of these must be supplied. Missing (zero-length) inputs are
treated as \code{NA} to allow proper Kleene OR logic.
}
\examples{
A(ab40 = "+")

A(ab40 = "+", ab42 = "-")

A(ab40 = "-", ab42 = "-")

}
