% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alz-package.R
\docType{package}
\name{alz-package}
\alias{alz}
\alias{alz-package}
\title{alz: Criteria for Diagnosis and Staging of Alzheimer's Disease}
\description{
Tools for representing, organizing, and analyzing diagnostic and staging information related to Alzheimer's disease. Supports structured workflows for combining clinical, biological, and other staging variables in a consistent and reproducible manner. Intended for use in research, data harmonization, and interpretation of disease progression frameworks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.pattern.institute/alz/}
  \item \url{https://github.com/patterninstitute/alz}
  \item Report bugs at \url{https://github.com/patterninstitute/alz/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Miguel Casanova \email{miguel.casanova@firalis.com} (\href{https://orcid.org/0000-0001-8622-6550}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute (https://www.pattern.institute) [copyright holder, funder]
}

}
\keyword{internal}
