% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_doc.R
\name{assay_doc}
\alias{assay_doc}
\title{Open assay reference document}
\usage{
assay_doc(assay_id)
}
\arguments{
\item{assay_id}{A character string representing the assay identifier (e.g.
\code{"innotest-ab42-csf"}). Check \code{\link[=assays]{assays()}} for options.}
}
\value{
Called for its side effect. Opens the PDF document in the system
viewer.
}
\description{
\code{assay_doc()} opens the technical fact sheet associated with a given assay.
This is useful for quickly checking specific assay implementations.
}
\examples{
\dontrun{
assay_doc("innotest-ab42-csf")
}

}
