% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atn.R
\name{ATN}
\alias{ATN}
\title{Perform ATN classification}
\usage{
ATN(A, T, N, classification = c("2018"))
}
\arguments{
\item{A}{A-status.}

\item{T}{T-status.}

\item{N}{N-status.}

\item{classification}{ATN classification scheme.}
}
\value{
To-be-documented.
}
\description{
At the moment, \code{\link[=ATN]{ATN()}} performs only the ATN 2018 scheme classification.
}
\examples{
(atn_grid <- atn_grid_2018())

ATN(A = atn_grid$A, T = atn_grid$T, N = atn_grid$N)

}
