% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atn_2018.R
\name{atn_grid_2018}
\alias{atn_grid_2018}
\title{Generate all 2018 ATN combinations}
\usage{
atn_grid_2018(include_na = FALSE)
}
\arguments{
\item{include_na}{Whether to include \code{NA} values.}
}
\value{
A tibble with columns \code{A}, \code{T} and \code{N}.
}
\description{
\code{\link[=atn_grid_2018]{atn_grid_2018()}} generates 2018 ATN combinations.
}
