% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_by.R
\name{bin_by}
\alias{bin_by}
\title{Categorize biomarker values into positivity bins}
\usage{
bin_by(x, cutpts)
}
\arguments{
\item{x}{A numeric vector of biomarker values.}

\item{cutpts}{A numeric vector of one or two strictly increasing
cut-points.}
}
\value{
A factor with levels \code{"-"}, \code{"+"}, or \code{"-"}, \code{"~"}, \code{"+"} depending
on the number of \code{cutpts}.
}
\description{
\code{\link[=bin_by]{bin_by()}} converts numeric biomarker values to diagnostic categories using
one or two cut-points. Values equal to a cut-point are assigned to the
\emph{higher} bin (i.e., positive or indeterminate), consistent with clinical
thresholding practices.
}
\examples{
# Using one cut-point
bin_by(x = 0:10, cutpts = 5)

# Using two cut-points
bin_by(x = 0:10, cutpts = c(5, 7))

# Using cutpoints from the curated table `cutpoints()`
bin_by(x = c(800, 1050, 1300), cutpts = cutpts("ab42-csf-elecsys-willemse2018"))

}
