% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmk.R
\name{bmk}
\alias{bmk}
\title{Biomarker constructor}
\usage{
bmk(x = character())
}
\arguments{
\item{x}{A character vector with values \code{"-"}, \code{"~"} or \code{"+"}.}
}
\value{
A vector classed \code{bmk}.
}
\description{
\code{\link[=bmk]{bmk()}} generates a biomarker-classed vector (\code{bmk}).
}
\examples{
bmk(c("-", "~", "+", NA))

}
