% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify biomarker values based on thresholds and quantification limits}
\usage{
classify(
  x,
  cutpoints,
  levels,
  loq = c(-Inf, Inf),
  right = TRUE,
  include.lowest = TRUE
)
}
\arguments{
\item{x}{Numeric vector of biomarker values.}

\item{cutpoints}{Numeric vector of cut-off points. Must be strictly increasing.}

\item{levels}{Character vector of classification labels, with length equal to \code{length(cutpoints) + 1}.}

\item{loq}{Numeric vector of length 2 specifying quantification limits:
lower (LLOQ) and upper (ULOQ). Default: \code{c(-Inf, Inf)}.
Optionally, you can name the elements to assign labels to values outside the range, e.g.:
\code{c("Below LLOQ" = 100, "Above ULOQ" = 2000)}.}

\item{right}{Logical. Intervals are right-closed (i.e. \verb{(a, b]}). Default: \code{TRUE}.}

\item{include.lowest}{Logical. Include the lowest value in the first interval. Default: \code{TRUE}.}
}
\value{
A factor with classification levels, including optional labels for values
outside the quantifiable range.
}
\description{
\code{\link[=classify]{classify()}} is a general-purpose function for discretising continuous
biomarker values into categorical levels based on user-defined cut-points.
It also handles values outside the quantifiable range, using optional
labels for values below the lower limit of quantification (LLOQ) or above
the upper limit (ULOQ), if provided via names in the \code{loq} argument.
}
\examples{
classify(
  x = c(10, 75, 130, 2050),
  cutpoints = c(50, 150),
  levels = c("low", "medium", "high"),
  loq = c("Below LLOQ" = 20, "Above ULOQ" = 2000)
)

}
