% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_by.R
\name{cut_by}
\alias{cut_by}
\title{Categorize biomarker values into positivity bins}
\usage{
cut_by(x, cutpts)
}
\arguments{
\item{x}{A numeric vector of biomarker values.}

\item{cutpts}{An object of class \code{cutpts}, see \code{\link[=cutpts]{cutpts()}} for more details.}
}
\value{
A factor with levels \code{"-"}, \code{"~"}, \code{"+"}.
}
\description{
\code{\link[=cut_by]{cut_by()}} converts numeric biomarker values to diagnostic categories using
one or two cut-points. Values equal to a cut-point are assigned to the
\emph{higher} bin (i.e., positive or indeterminate), consistent with clinical
thresholding practices.
}
\examples{
# Using one cut-point
cut_by(x = 0:10, cutpts = cutpts(5))

# Using two cut-points
cut_by(x = 0:10, cutpts = cutpts(c(5, 7)))

# Using cutpoints from the curated table `cutpoints()`
cut_by(x = c(800, 1050, 1300), cutpts = cutpts("ab42-csf-elecsys-willemse2018"))

}
