% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{cutpt_doc}
\alias{cutpt_doc}
\title{Open cut-point reference document}
\usage{
cutpt_doc(cutpt_id)
}
\arguments{
\item{cutpt_id}{A character string representing the assay identifier (e.g.
\code{"ab42-csf-elecsys-adni-upenn2013"}). Check \code{\link[=cutpoints]{cutpoints()}} for options.}
}
\value{
Called for its side effect. Opens the PDF document in the system
viewer.
}
\description{
\code{cutpt_doc()} opens the reference document describing the cut-point choice.
This is useful for quickly checking the documentation source for cut-points.
}
\examples{
\dontrun{
assay_doc("ab42-csf-elecsys-adni-upenn2013")
}

}
