% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loq.R
\name{loq}
\alias{loq}
\title{Limits of quantification per method}
\usage{
loq(assay_id)
}
\arguments{
\item{assay_id}{A method identifier. One of the values in the \code{assay_id}
variable in \code{\link[=assays]{assays()}}.}
}
\value{
A named numeric vector of two elements: \code{lloq}, lower, and \code{uloq},
upper, limits of quantification.
}
\description{
\code{\link[=loq]{loq()}} is a convenience function that retrieves the limits of quantification
from one of the assays in \code{\link[=assays]{assays()}}.
}
\examples{
# Limits of quantification of Fujirebio INNOTEST β-Amyloid(1-42) CSF ELISA
# Check `assays()` for available methods and associated identifiers.
loq(assay_id = "innotest-ab42-csf")

# Likewise, for Beta-Amyloid(1-40)
loq(assay_id = "innotest-ab40-csf")

}
