% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N}
\alias{N}
\title{Combine N biomarkers}
\usage{
N(nfl, mri_atrophy, fdg_pet, .classification = c("2018", "2024"))
}
\arguments{
\item{nfl}{bmk for Neurofilament light chain (optional)}

\item{mri_atrophy}{bmk for MRI-derived atrophy (optional)}

\item{fdg_pet}{bmk for FDG PET (optional)}

\item{.classification}{Either "2018" (default) or eventually "2024", etc.}
}
\value{
A \code{bmk} vector representing combined N-status under selected scheme.
}
\description{
Combine N biomarkers
}
\examples{
N(nfl = "+")

N(mri_atrophy = "+")

N(nfl = "+", mri_atrophy = "+")

}
