% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N2018}
\alias{N2018}
\title{N biomarkers (2018 scheme)}
\usage{
N2018(nfl, mri_atrophy, fdg_pet)
}
\arguments{
\item{nfl}{bmk vector for Neurofilament light chain (optional)}

\item{mri_atrophy}{bmk vector for MRI-derived atrophy (optional)}

\item{fdg_pet}{bmk vector for FDG PET (optional)}
}
\value{
A \code{bmk} vector representing the combined N-status.
}
\description{
Computes the overall N-status by OR-combining any subset of these biomarkers:
\itemize{
\item nfl
\item mri_atrophy
\item fdg_pet
At least one non-empty biomarker must be supplied. Zero-length inputs are ignored.
}
}
\keyword{internal}
