% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.R
\name{T}
\alias{T}
\title{Combine T biomarkers}
\usage{
T(
  t_tau,
  ptau217,
  ptau181,
  ptau231,
  ptau181_ab42_ratio,
  ptau217_np217_ratio,
  ptau205,
  mtbr_tau243,
  np_tau,
  tau_pet,
  ttau_ab42_ratio,
  .classification = c("2018", "2024")
)
}
\arguments{
\item{t_tau}{bmk for total tau (optional)}

\item{ptau217}{bmk for p-tau217 (optional)}

\item{ptau181}{bmk for p-tau181 (optional)}

\item{ptau231}{bmk for p-tau231 (optional)}

\item{ptau181_ab42_ratio}{bmk for p-tau181/AB42 ratio (optional)}

\item{ptau217_np217_ratio}{bmk for p-tau217/non-p-tau217 ratio (optional)}

\item{ptau205}{bmk for p-tau205 (optional)}

\item{mtbr_tau243}{bmk for MTBR-tau243 (optional)}

\item{np_tau}{bmk for non-phospho tau (optional)}

\item{tau_pet}{bmk for tau PET (optional)}

\item{ttau_ab42_ratio}{bmk for total tau/AB42 ratio (optional)}

\item{.classification}{Either "2018" (default) or eventually "2024", etc.}
}
\value{
A \code{bmk} vector representing combined T-status under selected scheme.
}
\description{
Combine T biomarkers
}
