% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.R
\name{T2018}
\alias{T2018}
\title{T biomarkers (2018 scheme)}
\usage{
T2018(
  t_tau,
  ptau217,
  ptau181,
  ptau231,
  ptau181_ab42_ratio,
  ptau217_np217_ratio,
  ptau205,
  mtbr_tau243,
  np_tau,
  tau_pet,
  ttau_ab42_ratio
)
}
\arguments{
\item{t_tau}{bmk vector for total tau (optional)}

\item{ptau217}{bmk vector for p-tau217 (optional)}

\item{ptau181}{bmk vector for p-tau181 (optional)}

\item{ptau231}{bmk vector for p-tau231 (optional)}

\item{ptau181_ab42_ratio}{bmk vector for p-tau181/AB42 ratio (optional)}

\item{ptau217_np217_ratio}{bmk vector for p-tau217/non-p-tau217 ratio (optional)}

\item{ptau205}{bmk vector for p-tau205 (optional)}

\item{mtbr_tau243}{bmk vector for MTBR-tau243 (optional)}

\item{np_tau}{bmk vector for non-phospho tau (optional)}

\item{tau_pet}{bmk vector for tau PET (optional)}

\item{ttau_ab42_ratio}{bmk vector for total tau/AB42 ratio (optional)}
}
\value{
A \code{bmk} vector representing the combined T-status.
}
\description{
Computes the overall T-status by OR-combining any subset of these biomarkers:
\itemize{
\item t_tau
\item ptau217
\item ptau181
\item ptau231
\item ptau181_ab42_ratio
\item ptau217_np217_ratio
\item ptau205
\item mtbr_tau243
\item np_tau
\item tau_pet
\item ttau_ab42_ratio
At least one non-empty biomarker must be supplied. Zero-length inputs are ignored.
}
}
\keyword{internal}
